grilo: Fix Guardian Eye Witness entries handling
The source had 2 bugs. First, it expected page sizes downloaded to be of
"count", but page sizes are always 20. After that, we didn't handle
left-over skip.

For example, when getting the entries from 75 to 100 (as grilo-test-ui
does), we need to get the 3rd page (from entries 61 to 80) and skip 15
items, followed by the 4th page (from entries 81 to 100).

See https://bugzilla.gnome.org/show_bug.cgi?id=764078